Revert "Merge branch 'tweakUIText_DriveDetectionWarning' of https://github.com/alekhyareddy28/PowerToys into tweakUIText_DriveDetectionWarning"
This reverts commit b193905d00dffc1c7ff5bdaaded461862c873128, reversing
changes made to 0d0db138f8d5e2551945312c08ed039d43d92784.
