+++ .gitignore
@@ -341,4 +341,3 @@ src/common/Telemetry/*.etl
 # Don't ignore MergeModules
 !**/MergeModules/Release/
 !**/MergeModules/Debug/
-/src/modules/previewpane/SvgThumbnailProvider/$(SolutionDir)$(Platform)/$(Configuration)/modules/FileExplorerPreview/SvgThumbnailProvider.xml
