+++ installer/PowerToysSetup/Product.wxs
@@ -428,23 +428,6 @@
           <RegistryValue Type="string" Key="InprocServer32\$(var.Version).0" Name="RuntimeVersion" Value="v4.0.30319" />
           <RegistryValue Type="string" Key="InprocServer32\$(var.Version).0" Name="CodeBase" Value="file:///[FileExplorerPreviewInstallFolder]SvgPreviewHandler.dll" />
         </RegistryKey>
-        <!-- Registry Key for Class Registration of Svg Thumbnail Provider -->
-        <RegistryKey Root="HKCR" Key="CLSID\{36B27788-A8BB-4698-A756-DF9F11F64F84}">
-            <RegistryValue Type="string" Value="SvgThumbnailProvider.SvgThumbnailProvider" />
-            <RegistryValue Type="string" Name="DisplayName" Value="Svg Thumbnail Provider" />
-            <RegistryValue Type="string" Name="AppID" Value="{CF142243-F059-45AF-8842-DBBE9783DB14}" />
-            <RegistryValue Type="string" Key="Implemented Categories\{62C8FE65-4EBB-45e7-B440-6E39B2CDBF29}" Value=""/>
-            <RegistryValue Type="string" Key="InprocServer32" Value="mscoree.dll" />
-            <RegistryValue Type="string" Key="InprocServer32" Name="Assembly" Value="SvgThumbnailProvider, Version=$(var.Version).0, Culture=neutral" />
-            <RegistryValue Type="string" Key="InprocServer32" Name="Class" Value="SvgThumbnailProvider.SvgThumbnailProvider" />
-            <RegistryValue Type="string" Key="InprocServer32" Name="RuntimeVersion" Value="v4.0.30319" />
-            <RegistryValue Type="string" Key="InprocServer32" Name="ThreadingModel" Value="Both" />
-            <RegistryValue Type="string" Key="InprocServer32" Name="CodeBase" Value="file:///[FileExplorerPreviewInstallFolder]SvgThumbnailProvider.dll" />
-            <RegistryValue Type="string" Key="InprocServer32\$(var.Version).0" Name="Assembly" Value="SvgThumbnailProvider, Version=$(var.Version).0, Culture=neutral" />
-            <RegistryValue Type="string" Key="InprocServer32\$(var.Version).0" Name="Class" Value="SvgThumbnailProvider.SvgThumbnailProvider" />
-            <RegistryValue Type="string" Key="InprocServer32\$(var.Version).0" Name="RuntimeVersion" Value="v4.0.30319" />
-            <RegistryValue Type="string" Key="InprocServer32\$(var.Version).0" Name="CodeBase" Value="file:///[FileExplorerPreviewInstallFolder]SvgThumbnailProvider.dll" />
-        </RegistryKey>
         <!-- Registry Key for Class Registration of Markdown Preview Handler -->
         <RegistryKey Root="HKCR" Key="CLSID\{45769bcc-e8fd-42d0-947e-02beef77a1f5}">
           <RegistryValue Type="string" Value="MarkdownPreviewHandler.MarkdownPreviewHandler" />
@@ -478,10 +461,6 @@
         <RegistryKey Root="HKCR" Key=".svg\shellex">
           <RegistryValue Type="string" Key="{8895b1c6-b41f-4c1c-a562-0d564250836f}" Value="{ddee2b8a-6807-48a6-bb20-2338174ff779}" />
         </RegistryKey>
-        <!-- Add file type association for Svg Thumbnail Provider -->
-        <RegistryKey Root="HKCR" Key=".svg\shellex">
-            <RegistryValue Type="string" Key="{E357FCCD-A995-4576-B01F-234630154E96}" Value="{36B27788-A8BB-4698-A756-DF9F11F64F84}" />
-        </RegistryKey>
         <!-- Add file type association for Markdown Preview Handler -->
         <RegistryKey Root="HKCR" Key=".md\shellex">
           <RegistryValue Type="string" Key="{8895b1c6-b41f-4c1c-a562-0d564250836f}" Value="{45769bcc-e8fd-42d0-947e-02beef77a1f5}" />
@@ -489,10 +468,6 @@
         <!-- Update Key to use IE11 for prevhost.exe -->
         <RegistryKey Root="HKLM" Key="Software\Microsoft\Internet Explorer\Main\FeatureControl\FEATURE_BROWSER_EMULATION">
           <RegistryValue Type="integer" Name="prevhost.exe" Value="11000" />
-        </RegistryKey>
-        <!-- Update Key to use IE11 for dllhost.exe -->
-        <RegistryKey Root="HKLM" Key="Software\Microsoft\Internet Explorer\Main\FeatureControl\FEATURE_BROWSER_EMULATION">
-            <RegistryValue Type="integer" Name="dllhost.exe" Value="11000" />
         </RegistryKey>
       </Component>
     </DirectoryRef>
@@ -636,8 +611,6 @@
         <File Source="$(var.BinX64Dir)modules\FileExplorerPreview\Telemetry.dll" />
         <!-- File to include dll for Svg Preview Handler -->
         <File Source="$(var.BinX64Dir)modules\FileExplorerPreview\SvgPreviewHandler.dll" />
-        <!-- File to include dll for Svg Preview Handler -->
-        <File Source="$(var.BinX64Dir)modules\FileExplorerPreview\SvgThumbnailProvider.dll" />
         <!-- Files to include dll's for Markdown Preview Handler and it's dependencies -->
         <File Source="$(var.BinX64Dir)modules\FileExplorerPreview\MarkdownPreviewHandler.dll" />
         <File Source="$(var.BinX64Dir)modules\FileExplorerPreview\Markdig.Signed.dll" />
