+++ src/core/Microsoft.PowerToys.Settings.UI/ViewModels/PowerPreviewViewModel.cs
@@ -27,14 +27,12 @@ namespace Microsoft.PowerToys.Settings.UI.ViewModels
                 SettingsUtils.SaveSettings(Settings.ToJsonString(), ModuleName);
             }
 
-            this._svgRenderIsEnabled = Settings.properties.EnableSvgPreview;
-            this._svgThumbnailIsEnabled = Settings.properties.EnableSvgThumbnail;
-            this._mdRenderIsEnabled = Settings.properties.EnableMdPreview;
+            this._svgRenderIsEnabled = Settings.properties.EnableSvg;
+            this._mdRenderIsEnabled = Settings.properties.EnableMd;
         }
 
         private bool _svgRenderIsEnabled = false;
         private bool _mdRenderIsEnabled = false;
-        private bool _svgThumbnailIsEnabled = false;
 
         public bool SVGRenderIsEnabled
         {
@@ -48,25 +46,7 @@ namespace Microsoft.PowerToys.Settings.UI.ViewModels
                 if (value != _svgRenderIsEnabled)
                 {
                     _svgRenderIsEnabled = value;
-                    Settings.properties.EnableSvgPreview = value;
-                    RaisePropertyChanged();
-                }
-            }
-        }
-
-        public bool SVGThumbnailIsEnabled
-        {
-            get
-            {
-                return _svgThumbnailIsEnabled;
-            }
-
-            set
-            {
-                if (value != _svgThumbnailIsEnabled)
-                {
-                    _svgThumbnailIsEnabled = value;
-                    Settings.properties.EnableSvgThumbnail = value;
+                    Settings.properties.EnableSvg = value;
                     RaisePropertyChanged();
                 }
             }
@@ -84,7 +64,7 @@ namespace Microsoft.PowerToys.Settings.UI.ViewModels
                 if (value != _mdRenderIsEnabled)
                 {
                     _mdRenderIsEnabled = value;
-                    Settings.properties.EnableMdPreview = value;
+                    Settings.properties.EnableMd = value;
                     RaisePropertyChanged();
                 }
             }
