+++ src/core/Microsoft.PowerToys.Settings.UnitTest/ViewModelTests/PowerPreview.cs
@@ -59,28 +59,11 @@ namespace ViewModelTests
             ShellPage.DefaultSndMSGCallback = msg =>
             {
                 SndModuleSettings<SndPowerPreviewSettings> snd = JsonSerializer.Deserialize<SndModuleSettings<SndPowerPreviewSettings>>(msg);
-                Assert.IsTrue(snd.powertoys.FileExplorerPreviewSettings.properties.EnableSvgPreview);
+                Assert.IsTrue(snd.powertoys.FileExplorerPreviewSettings.properties.EnableSvg);
             };
 
             // act
             viewModel.SVGRenderIsEnabled = true;
-        }
-
-        [TestMethod]
-        public void SVGThumbnailIsEnabled_ShouldPrevHandler_WhenSuccessful()
-        {
-            // arrange
-            PowerPreviewViewModel viewModel = new PowerPreviewViewModel();
-
-            // Assert
-            ShellPage.DefaultSndMSGCallback = msg =>
-            {
-                SndModuleSettings<SndPowerPreviewSettings> snd = JsonSerializer.Deserialize<SndModuleSettings<SndPowerPreviewSettings>>(msg);
-                Assert.IsTrue(snd.powertoys.FileExplorerPreviewSettings.properties.EnableSvgThumbnail);
-            };
-
-            // act
-            viewModel.SVGThumbnailIsEnabled = true;
         }
 
         [TestMethod]
@@ -93,7 +76,7 @@ namespace ViewModelTests
             ShellPage.DefaultSndMSGCallback = msg =>
             {
                 SndModuleSettings<SndPowerPreviewSettings> snd = JsonSerializer.Deserialize<SndModuleSettings<SndPowerPreviewSettings>>(msg);
-                Assert.IsTrue(snd.powertoys.FileExplorerPreviewSettings.properties.EnableMdPreview);
+                Assert.IsTrue(snd.powertoys.FileExplorerPreviewSettings.properties.EnableMd);
             };
 
             // act
