--- src/modules/previewpane/UnitTests-SvgThumbnailProvider/packages.config
@@ -1,9 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Castle.Core" version="4.4.0" targetFramework="net472" />
-  <package id="Moq" version="4.14.5" targetFramework="net472" />
-  <package id="MSTest.TestAdapter" version="2.1.1" targetFramework="net472" />
-  <package id="MSTest.TestFramework" version="2.1.1" targetFramework="net472" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.0" targetFramework="net472" />
-  <package id="System.Threading.Tasks.Extensions" version="4.5.1" targetFramework="net472" />
-</packages>
