+++ src/modules/previewpane/common/PreviewHandlerCommon.csproj
@@ -103,8 +103,6 @@
     <Compile Include="cominterop\IPreviewHandler.cs" />
     <Compile Include="cominterop\IPreviewHandlerFrame.cs" />
     <Compile Include="cominterop\IPreviewHandlerVisuals.cs" />
-    <Compile Include="cominterop\IThumbnailProvider.cs" />
-    <Compile Include="cominterop\IViewObject.cs" />
     <Compile Include="cominterop\LOGFONT.cs" />
     <Compile Include="cominterop\MSG.cs" />
     <Compile Include="cominterop\RECT.cs" />
@@ -127,7 +125,6 @@
     <Compile Include="handlers\StreamBasedPreviewHandler.cs" />
     <Compile Include="examplehandler\TestCustomHandler.cs" />
     <Compile Include="Utilities\StreamWrapper.cs" />
-    <Compile Include="Utilities\SvgPreviewHandlerHelper.cs" />
   </ItemGroup>
   <ItemGroup>
     <AdditionalFiles Include="..\..\..\codeAnalysis\StyleCop.json">
