--- src/modules/previewpane/common/cominterop/IThumbnailProvider.cs
@@ -1,47 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation
-// The Microsoft Corporation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System;
-using System.Runtime.InteropServices;
-
-namespace Common.ComInterlop
-{
-    /// <summary>
-    /// Specifies the alpha type of the image.
-    /// </summary>
-    public enum WTS_ALPHATYPE : int
-    {
-        /// <summary>
-        /// he bitmap is an unknown format. The Shell tries nonetheless to detect whether the image has an alpha channel.
-        /// </summary>
-        WTSAT_UNKNOWN = 0,
-
-        /// <summary>
-        /// The bitmap is an RGB image without alpha. The alpha channel is invalid and the Shell ignores it.
-        /// </summary>
-        WTSAT_RGB = 1,
-
-        /// <summary>
-        /// The bitmap is an ARGB image with a valid alpha channel.
-        /// </summary>
-        WTSAT_ARGB = 2,
-    }
-
-    /// <summary>
-    /// Exposes methods for thumbnail provider.
-    /// </summary>
-    [ComImport]
-    [Guid("e357fccd-a995-4576-b01f-234630154e96")]
-    [InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
-    public interface IThumbnailProvider
-    {
-        /// <summary>
-        /// Gets a thumbnail image and alpha type.
-        /// </summary>
-        /// <param name="cx">The maximum thumbnail size, in pixels.</param>
-        /// <param name="phbmp">When this method returns, contains a pointer to the thumbnail image handle.</param>
-        /// <param name="pdwAlpha">When this method returns, contains a pointer to one of the values from the WTS_ALPHATYPE enumeration.</param>
-        void GetThumbnail(uint cx, out IntPtr phbmp, out WTS_ALPHATYPE pdwAlpha);
-    }
-}
