+++ src/modules/previewpane/powerpreview/CLSID.h
@@ -16,15 +16,8 @@ const CLSID CLSID_SHIMActivateMdPreviewHandler = { 0xE0907A95, 0x6F9A, 0x4D1B, {
 // 45769bcc-e8fd-42d0-947e-02beef77a1f5
 const CLSID CLSID_MdPreviewHandler = { 0x45769bcc, 0xe8fd, 0x42d0, { 0x94, 0x7e, 0x02, 0xbe, 0xef, 0x77, 0xa1, 0xf5 } };
 
-// 9C723B8C-4F5C-4147-9DE4-C2808F9AF66B
-const CLSID CLSID_SHIMActivateSvgThumbnailProvider = { 0x9C723B8C, 0x4F5C, 0x4147, { 0x9D, 0xE4, 0xC2, 0x80, 0x8F, 0x9A, 0xF6, 0x6B } };
-
-// 36B27788-A8BB-4698-A756-DF9F11F64F84
-const CLSID CLSID_SvgThumbnailProvider = { 0x36B27788, 0xA8BB, 0x4698, { 0xA7, 0x56, 0xDF, 0x9F, 0x11, 0xF6, 0x4F, 0x84 } };
-
 // Pairs of NativeClsid vs ManagedClsid used for preview handlers.
 const std::vector<std::pair<CLSID, CLSID>> NativeToManagedClsid({
     { CLSID_SHIMActivateMdPreviewHandler, CLSID_MdPreviewHandler },
-    { CLSID_SHIMActivateSvgPreviewHandler, CLSID_SvgPreviewHandler },
-    { CLSID_SHIMActivateSvgThumbnailProvider, CLSID_SvgThumbnailProvider }
+    { CLSID_SHIMActivateSvgPreviewHandler, CLSID_SvgPreviewHandler }
 });
