+++ src/modules/previewpane/powerpreview/powerpreview.cpp
@@ -26,20 +26,6 @@ void PowerPreviewModule::destroy()
         }
     }
 
-    for (auto thumbnailProvider : this->m_thumbnailProviders)
-    {
-        if (thumbnailProvider != NULL)
-        {
-            // Disable all the active thumbnail providers.
-            if (this->m_enabled && thumbnailProvider->GetToggleSettingState())
-            {
-                thumbnailProvider->DisablePreview();
-            }
-
-            delete thumbnailProvider;
-        }
-    }
-
     delete this;
 }
 
@@ -81,15 +67,6 @@ bool PowerPreviewModule::get_config(_Out_ wchar_t* buffer, _Out_ int* buffer_siz
             previewHandler->GetToggleSettingState());
     }
 
-    for (auto thumbnailProvider : this->m_thumbnailProviders)
-    {
-        settings.add_bool_toggle(
-            thumbnailProvider->GetToggleSettingName(),
-            thumbnailProvider->GetToggleSettingDescription(),
-            thumbnailProvider->GetToggleSettingState());
-    }
-
-
     return settings.serialize_to_buffer(buffer, buffer_size);
 }
 
@@ -103,11 +80,6 @@ void PowerPreviewModule::set_config(const wchar_t* config)
         for (auto previewHandler : this->m_previewHandlers)
         {
             previewHandler->UpdateState(settings, this->m_enabled);
-        }
-
-        for (auto thumbnailProvider : this->m_thumbnailProviders)
-        {
-            thumbnailProvider->UpdateState(settings, this->m_enabled);
         }
 
         settings.save_to_settings_file();
@@ -134,19 +106,6 @@ void PowerPreviewModule::enable()
         }
     }
 
-    for (auto thumbnailProvider : this->m_thumbnailProviders)
-    {
-        if (thumbnailProvider->GetToggleSettingState())
-        {
-            // Enable all the thumbnail providers with initial state set as true.
-            thumbnailProvider->EnableThumbnailProvider();
-        }
-        else
-        {
-            thumbnailProvider->DisableThumbnailProvider();
-        }
-    }
-
     if (!this->m_enabled)
     {
         Trace::EnabledPowerPreview(true);
@@ -161,11 +120,6 @@ void PowerPreviewModule::disable()
     for (auto previewHandler : this->m_previewHandlers)
     {
         previewHandler->DisablePreview();
-    }
-
-    for (auto thumbnailProvider : this->m_thumbnailProviders)
-    {
-        thumbnailProvider->DisableThumbnailProvider();
     }
 
     if (this->m_enabled)
@@ -202,12 +156,6 @@ void PowerPreviewModule::init_settings()
         {
             previewHandler->LoadState(settings);
         }
-
-        for (auto thumbnailProvider : this->m_thumbnailProviders)
-        {
-            thumbnailProvider->LoadState(settings);
-        }
-
     }
     catch (std::exception const& e)
     {
