+++ src/modules/previewpane/powerpreview/powerpreview.h
@@ -17,7 +17,6 @@ private:
     bool m_enabled = false;
     std::wstring m_moduleName;
     std::vector<FileExplorerPreviewSettings *> m_previewHandlers;
-    std::vector<FileExplorerPreviewSettings*> m_thumbnailProviders;
 
 public:
     PowerPreviewModule() :
@@ -39,16 +38,6 @@ public:
                   GET_RESOURCE_STRING(IDS_PREVPANE_MD_SETTINGS_DESCRIPTION),
                   L"{45769bcc-e8fd-42d0-947e-02beef77a1f5}",
                   L"Markdown Preview Handler",
-                  new RegistryWrapper())
-            }),
-        m_thumbnailProviders(
-            { // TODO: MOVE THIS SVG Thumbnail Provider settings object.
-              new FileExplorerPreviewSettings(
-                  true,
-                  L"svg-thumbnail-toggle-setting",
-                  GET_RESOURCE_STRING(IDS_SVG_THUMBNAIL_PROVIDER_SETTINGS_DESCRIPTION),
-                  L"{36B27788-A8BB-4698-A756-DF9F11F64F84}",
-                  L"SVG Thumbnail Provider",
                   new RegistryWrapper())
             })
     {
