+++ src/modules/previewpane/powerpreview/powerpreview.rc
@@ -95,7 +95,6 @@ BEGIN
     IDS_PREVPANE_SVG_BOOL_TOGGLE_CONTROLL               L"IDS_PREVPANE_SVG_BOOL_TOGGLE_CONTROLL"
     IDS_PREVPANE_SVG_SETTINGS_DESCRIPTION               L"Svg Previewer"
     IDS_PREVPANE_SVG_SETTINGS_DISPLAYNAME               L"SVG Previewer"
-    IDS_SVG_THUMBNAIL_PROVIDER_SETTINGS_DESCRIPTION     L"Svg Thumbnail Provider"
     IDS_EXPLR_SVG_BOOL_TOGGLE_CONTROLL                  L"EXPLR_SVG_BOOL_TOGGLE_CONTROLL"
     IDS_EXPLR_SVG_SETTINGS_DESCRIPTION                  L"Render SVG images"
 END
