+++ src/modules/previewpane/powerpreview/settings.h
@@ -7,24 +7,24 @@
 
 namespace PowerPreviewSettings
 {
-    // PowerToy Windows Explorer File Preview Settings.
-    class FileExplorerPreviewSettings
-    {
-    private:
-        bool m_toggleSettingEnabled;
+	// PowerToy Windows Explore File Preview Settings.
+	class FileExplorerPreviewSettings
+	{
+	private:
+		bool m_toggleSettingEnabled;
         std::wstring m_toggleSettingName;
         std::wstring m_toggleSettingDescription;
-        std::wstring m_registryValueData;
+		std::wstring m_registryValueData;
         RegistryWrapperIface * m_registryWrapper;
-        LPCWSTR m_clsid;
+		LPCWSTR m_clsid;
 
-    public:
+	public:
         FileExplorerPreviewSettings(bool toggleSettingEnabled, const std::wstring& toggleSettingName, const std::wstring& toggleSettingDescription, LPCWSTR clsid, const std::wstring& registryValueData, RegistryWrapperIface* registryWrapper);
         ~ FileExplorerPreviewSettings();
 
-        virtual bool GetToggleSettingState() const;
+		virtual bool GetToggleSettingState() const;
         virtual void UpdateToggleSettingState(bool state);
-        virtual std::wstring GetToggleSettingName() const;
+		virtual std::wstring GetToggleSettingName() const;
         virtual std::wstring GetToggleSettingDescription() const;
         virtual LPCWSTR GetCLSID() const;
         virtual std::wstring GetRegistryValueData() const;
@@ -32,7 +32,5 @@ namespace PowerPreviewSettings
         virtual void UpdateState(PowerToysSettings::PowerToyValues& settings, bool enabled);
         virtual LONG EnablePreview();
         virtual LONG DisablePreview();
-        virtual LONG EnableThumbnailProvider();
-        virtual LONG DisableThumbnailProvider();
-    };
+	};
 }
